refactor(BA-6110): split deploying handlers into per-file modules
Move DeployingProvisioningHandler and DeployingRollingBackHandler into
their own modules and document the "1 handler = 1 file" rule under
sokovan/CLAUDE.md. No behavior change.
